projekat/requirements.txt
requests
redis
python-consul
Flask
python-dotenv
Flask-JWT-Extended